Get series of a dataset

This function returns a data frame with all series for a dataset

get_series(dataset)

Arguments

dataset

the slug of the dataset

Examples

get_series(dataset = "insee-act-trim-anc")

#>    attributes.decimals attributes.freq attributes.last-update
#> 1                    1               t             2017-02-16
#> 2                    1               t             2017-02-16
#> 3                    1               t             2017-02-16
#> 4                    1               t             2017-02-16
#> 5                    1               t             2017-02-16
#> 6                    1               t             2017-02-16
#> 7                    1               t             2017-02-16
#> 8                    1               t             2017-02-16
#> 9                    1               t             2017-02-16
#> 10                   1               t             2017-02-16
#> 11                   1               t             2017-02-16
#> 12                   1               t             2017-02-16
#> 13                   1               t             2017-02-16
#> 14                   1               t             2017-02-16
#> 15                   1               t             2017-02-16
#> 16                   1               t             2017-02-16
#> 17                   1               t             2017-02-16
#> 18                   1               t             2017-02-16
#> 19                   1               t             2017-02-16
#> 20                   1               t             2017-02-16
#>    attributes.ref-area attributes.time-per-collect
#> 1                   fm                     periode
#> 2                   fm                     periode
#> 3                   fm                     periode
#> 4                   fm                     periode
#> 5                   fm                     periode
#> 6                   fm                     periode
#> 7                   fm                     periode
#> 8                   fm                     periode
#> 9                   fm                     periode
#> 10                  fm                     periode
#> 11                  fm                     periode
#> 12                  fm                     periode
#> 13                  fm                     periode
#> 14                  fm                     periode
#> 15                  fm                     periode
#> 16                  fm                     periode
#> 17                  fm                     periode
#> 18                  fm                     periode
#> 19                  fm                     periode
#> 20                  fm                     periode
#>                                                                            attributes.title
#> 1           esperance-apparente-d-activite-au-sens-du-bit-femmes-de-50-a-64-ans-donnees-cvs
#> 2           esperance-apparente-d-activite-au-sens-du-bit-hommes-de-50-a-64-ans-donnees-cvs
#> 3        esperance-apparente-d-activite-au-sens-du-bit-ensemble-des-50-a-64-ans-donnees-cvs
#> 4           esperance-apparente-d-activite-au-sens-du-bit-femmes-de-55-a-64-ans-donnees-cvs
#> 5           esperance-apparente-d-activite-au-sens-du-bit-hommes-de-55-a-64-ans-donnees-cvs
#> 6        esperance-apparente-d-activite-au-sens-du-bit-ensemble-des-55-a-64-ans-donnees-cvs
#> 7     personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-15-a-24-ans-donnees-cvs
#> 8     personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-15-a-64-ans-donnees-cvs
#> 9     personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-25-a-49-ans-donnees-cvs
#> 10    personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-50-a-64-ans-donnees-cvs
#> 11 personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-50-ans-ou-plus-donnees-cvs
#> 12    personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-55-a-64-ans-donnees-cvs
#> 13                             personnes-actives-taux-d-activite-au-sens-du-bit-donnees-cvs
#> 14       personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-15-a-24-ans-donnees-cvs
#> 15       personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-25-a-49-ans-donnees-cvs
#> 16       personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-50-a-64-ans-donnees-cvs
#> 17       personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-55-a-64-ans-donnees-cvs
#> 18                      personnes-actives-taux-d-activite-au-sens-du-bit-femmes-donnees-cvs
#> 19       personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-15-a-64-ans-donnees-cvs
#> 20       personnes-actives-taux-d-activite-au-sens-du-bit-hommes-de-15-a-24-ans-donnees-cvs
#>    attributes.unit-measure attributes.unit-mult dataset_code dimensions.age
#> 1                       an                    0 ACT-TRIM-ANC          50-64
#> 2                       an                    0 ACT-TRIM-ANC          50-64
#> 3                       an                    0 ACT-TRIM-ANC          50-64
#> 4                       an                    0 ACT-TRIM-ANC          55-64
#> 5                       an                    0 ACT-TRIM-ANC          55-64
#> 6                       an                    0 ACT-TRIM-ANC          55-64
#> 7                      pct                    0 ACT-TRIM-ANC          15-24
#> 8                      pct                    0 ACT-TRIM-ANC          15-64
#> 9                      pct                    0 ACT-TRIM-ANC          25-49
#> 10                     pct                    0 ACT-TRIM-ANC          50-64
#> 11                     pct                    0 ACT-TRIM-ANC             50
#> 12                     pct                    0 ACT-TRIM-ANC      dont55-64
#> 13                     pct                    0 ACT-TRIM-ANC             00
#> 14                     pct                    0 ACT-TRIM-ANC          15-24
#> 15                     pct                    0 ACT-TRIM-ANC          25-49
#> 16                     pct                    0 ACT-TRIM-ANC          50-64
#> 17                     pct                    0 ACT-TRIM-ANC          55-64
#> 18                     pct                    0 ACT-TRIM-ANC             00
#> 19                     pct                    0 ACT-TRIM-ANC          15-64
#> 20                     pct                    0 ACT-TRIM-ANC          15-24
#>    dimensions.indicateur dimensions.sexe end_date                    end_ts
#> 1                     ea               2      187 2016-10-01T00:00:00+00:00
#> 2                     ea               1      187 2016-10-01T00:00:00+00:00
#> 3                     ea               0      187 2016-10-01T00:00:00+00:00
#> 4                     ea               2      187 2016-10-01T00:00:00+00:00
#> 5                     ea               1      187 2016-10-01T00:00:00+00:00
#> 6                     ea               0      187 2016-10-01T00:00:00+00:00
#> 7                     20               0      187 2016-10-01T00:00:00+00:00
#> 8                     20               0      187 2016-10-01T00:00:00+00:00
#> 9                     20               0      187 2016-10-01T00:00:00+00:00
#> 10                    20               0      187 2016-10-01T00:00:00+00:00
#> 11                    20               0      187 2016-10-01T00:00:00+00:00
#> 12                    20               0      187 2016-10-01T00:00:00+00:00
#> 13                    20               0      187 2016-10-01T00:00:00+00:00
#> 14                    20               2      187 2016-10-01T00:00:00+00:00
#> 15                    20               2      187 2016-10-01T00:00:00+00:00
#> 16                    20               2      187 2016-10-01T00:00:00+00:00
#> 17                    20               2      187 2016-10-01T00:00:00+00:00
#> 18                    20               2      187 2016-10-01T00:00:00+00:00
#> 19                    20               2      187 2016-10-01T00:00:00+00:00
#> 20                    20               1      187 2016-10-01T00:00:00+00:00
#>    frequency       key            last_update_ds          last_update_widu
#> 1          Q 001688389 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 2          Q 001688390 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 3          Q 001688391 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 4          Q 001688392 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 5          Q 001688393 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 6          Q 001688394 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 7          Q 001688407 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 8          Q 001688408 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 9          Q 001688409 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 10         Q 001688410 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 11         Q 001688411 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 12         Q 001688412 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 13         Q 001688413 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 14         Q 001688414 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 15         Q 001688415 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 16         Q 001688416 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 17         Q 001688417 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 18         Q 001688418 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 19         Q 001688419 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#> 20         Q 001688420 2017-02-16T00:00:00+00:00 2017-02-18T08:40:27+00:00
#>                                                        name provider_name
#> 1  Active life expectancy - Women - From 50 to 64 years old         INSEE
#> 2    Active life expectancy - Men - From 50 to 64 years old         INSEE
#> 3    Active life expectancy - All - From 50 to 64 years old         INSEE
#> 4  Active life expectancy - Women - From 55 to 64 years old         INSEE
#> 5    Active life expectancy - Men - From 55 to 64 years old         INSEE
#> 6    Active life expectancy - All - From 55 to 64 years old         INSEE
#> 7             Activity rate - All - From 15 to 24 years old         INSEE
#> 8             Activity rate - All - From 15 to 64 years old         INSEE
#> 9             Activity rate - All - From 25 to 49 years old         INSEE
#> 10            Activity rate - All - From 50 to 64 years old         INSEE
#> 11              Activity rate - All - 50 years old and over         INSEE
#> 12   Activity rate - All - Of which from 55 to 64 years old         INSEE
#> 13                           Activity rate - All - All ages         INSEE
#> 14          Activity rate - Women - From 15 to 24 years old         INSEE
#> 15          Activity rate - Women - From 25 to 49 years old         INSEE
#> 16          Activity rate - Women - From 50 to 64 years old         INSEE
#> 17          Activity rate - Women - From 55 to 64 years old         INSEE
#> 18                         Activity rate - Women - All ages         INSEE
#> 19          Activity rate - Women - From 15 to 64 years old         INSEE
#> 20            Activity rate - Men - From 15 to 24 years old         INSEE
#>                            slug start_date                  start_ts
#> 1  insee-act-trim-anc-001688389        132 2003-01-01T00:00:00+00:00
#> 2  insee-act-trim-anc-001688390        132 2003-01-01T00:00:00+00:00
#> 3  insee-act-trim-anc-001688391        132 2003-01-01T00:00:00+00:00
#> 4  insee-act-trim-anc-001688392        132 2003-01-01T00:00:00+00:00
#> 5  insee-act-trim-anc-001688393        132 2003-01-01T00:00:00+00:00
#> 6  insee-act-trim-anc-001688394        132 2003-01-01T00:00:00+00:00
#> 7  insee-act-trim-anc-001688407        132 2003-01-01T00:00:00+00:00
#> 8  insee-act-trim-anc-001688408        132 2003-01-01T00:00:00+00:00
#> 9  insee-act-trim-anc-001688409        132 2003-01-01T00:00:00+00:00
#> 10 insee-act-trim-anc-001688410        132 2003-01-01T00:00:00+00:00
#> 11 insee-act-trim-anc-001688411        132 2003-01-01T00:00:00+00:00
#> 12 insee-act-trim-anc-001688412        132 2003-01-01T00:00:00+00:00
#> 13 insee-act-trim-anc-001688413        132 2003-01-01T00:00:00+00:00
#> 14 insee-act-trim-anc-001688414        132 2003-01-01T00:00:00+00:00
#> 15 insee-act-trim-anc-001688415        132 2003-01-01T00:00:00+00:00
#> 16 insee-act-trim-anc-001688416        132 2003-01-01T00:00:00+00:00
#> 17 insee-act-trim-anc-001688417        132 2003-01-01T00:00:00+00:00
#> 18 insee-act-trim-anc-001688418        132 2003-01-01T00:00:00+00:00
#> 19 insee-act-trim-anc-001688419        132 2003-01-01T00:00:00+00:00
#> 20 insee-act-trim-anc-001688420        132 2003-01-01T00:00:00+00:00
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                              tags
#> 1  2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-femmes-de-50-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 2    2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-hommes-de-50-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, men, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 3      2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-ensemble-des-50-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 4  2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-femmes-de-55-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 5    2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-hommes-de-55-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, men, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 6      2016-03-03, act-trim-anc, active, activity, age, area, bdm, collection, date, decimals, economic, esperance-apparente-d-activite-au-sens-du-bit-ensemble-des-55-a-64-ans-donnees-cvs, expectancy, france, frequency, gender, group, identifier, indicators, insee, institute, last, life, metropolitan, multiplier, national, numerical, observation, old, period, quarterly, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 7                       2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-15-a-24-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 8                       2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-15-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 9                       2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-25-a-49-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 10                      2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-50-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 11                   2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-50-ans-ou-plus-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 12                      2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-ensemble-des-55-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#> 13                                                     2016-03-03, act-trim-anc, activity, age, ages, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, period, personnes-actives-taux-d-activite-au-sens-du-bit-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website
#> 14                  2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-15-a-24-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 15                  2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-25-a-49-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 16                  2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-50-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 17                  2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-55-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 18                                       2016-03-03, act-trim-anc, activity, age, ages, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women
#> 19                  2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-femmes-de-15-a-64-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, women, years
#> 20                    2016-03-03, act-trim-anc, activity, age, area, bdm, collection, date, decimals, economic, france, frequency, gender, group, identifier, indicators, insee, institute, last, men, metropolitan, multiplier, national, numerical, observation, old, period, personnes-actives-taux-d-activite-au-sens-du-bit-hommes-de-15-a-24-ans-donnees-cvs, quarterly, rate, reference, series, sex, statistics, studies, time, title, unit, update, used, website, years
#>    version
#> 1        3
#> 2        3
#> 3        3
#> 4        3
#> 5        3
#> 6        3
#> 7        3
#> 8        3
#> 9        3
#> 10       3
#> 11       3
#> 12       3
#> 13       3
#> 14       3
#> 15       3
#> 16       3
#> 17       3
#> 18       3
#> 19       3
#> 20       3